Add task role count
